+++ packages/core/src/indexing/postgres/migrations/1-create-model-table.ts
@@ -7,7 +7,8 @@ export async function createModelTable(dataSource: Knex, tableName: string) {
 
     table.string('stream_id').primary(`idx_${indexName}_pkey`).unique(`constr_${indexName}_unique`)
     table.string('controller_did', 1024).notNullable()
-    table.dateTime('last_anchored_at')
+    table.dateTime('last_anchored_at').nullable()
+    table.dateTime('first_anchored_at').nullable()
     table.dateTime('created_at').notNullable().defaultTo(dataSource.fn.now())
     table.dateTime('updated_at').notNullable().defaultTo(dataSource.fn.now())
 
@@ -15,6 +16,9 @@ export async function createModelTable(dataSource: Knex, tableName: string) {
       storageEngineIndexType: 'hash',
     })
     table.index(['last_anchored_at'], `idx_${indexName}_last_anchored_at`, {
+      storageEngineIndexType: 'hash',
+    })
+    table.index(['first_anchored_at'], `idx_${indexName}_first_anchored_at`, {
       storageEngineIndexType: 'hash',
     })
     table.index(['created_at'], `idx_${indexName}_created_at`, {