+++ packages/common/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0-rc.1](/compare/@ceramicnetwork/common@2.5.0-rc.0...@ceramicnetwork/common@2.5.0-rc.1) (2022-08-11)
+
+
+### Features
+
+* **stream-model-instance,stream-model-instance-handler:** Enable deterministic MIDs for models with SINGLE accountRelations (#2356) a39d5e2, closes #2356
+
+
+
+
+
+# [2.5.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/common@2.4.0...@ceramicnetwork/common@2.5.0-rc.0) (2022-08-10)
+
+
+### Features
+
+* **core:** working implementation of indexable anchors Phase 2 ([#2315](https://github.com/ceramicnetwork/js-ceramic/issues/2315)) ([987cd43](https://github.com/ceramicnetwork/js-ceramic/commit/987cd43fa5d6f0a8bac1aefc28e8b181e33b62cb))
+
+
+
+
+
+# [2.4.0](/compare/@ceramicnetwork/common@2.4.0-rc.1...@ceramicnetwork/common@2.4.0) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/common
+
+
+
+
+
 # [2.4.0-rc.1](/compare/@ceramicnetwork/common@2.4.0-rc.0...@ceramicnetwork/common@2.4.0-rc.1) (2022-07-26)
 
 **Note:** Version bump only for package @ceramicnetwork/common