+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.7-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/logger@2.0.6...@ceramicnetwork/logger@2.0.7-rc.0) (2022-08-10)
+
+**Note:** Version bump only for package @ceramicnetwork/logger
+
+
+
+
+
+## [2.0.6](/compare/@ceramicnetwork/logger@2.0.6-rc.1...@ceramicnetwork/logger@2.0.6) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/logger
+
+
+
+
+
 ## [2.0.6-rc.1](/compare/@ceramicnetwork/logger@2.0.6-rc.0...@ceramicnetwork/logger@2.0.6-rc.1) (2022-07-14)
 
 **Note:** Version bump only for package @ceramicnetwork/logger