+++ packages/stream-model-instance/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0-rc.0](/compare/@ceramicnetwork/stream-model-instance@0.3.2-rc.0...@ceramicnetwork/stream-model-instance@0.4.0-rc.0) (2022-08-11)
+
+
+### Features
+
+* **stream-model-instance,stream-model-instance-handler:** Enable deterministic MIDs for models with SINGLE accountRelations (#2356) a39d5e2, closes #2356
+
+
+
+
+
+## [0.3.2-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-model-instance@0.3.1...@ceramicnetwork/stream-model-instance@0.3.2-rc.0) (2022-08-10)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance
+
+
+
+
+
+## [0.3.1](/compare/@ceramicnetwork/stream-model-instance@0.3.1-rc.2...@ceramicnetwork/stream-model-instance@0.3.1) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance
+
+
+
+
+
 ## [0.3.1-rc.2](/compare/@ceramicnetwork/stream-model-instance@0.3.1-rc.1...@ceramicnetwork/stream-model-instance@0.3.1-rc.2) (2022-07-26)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance