+++ packages/stream-tile-handler/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5-rc.1](/compare/@ceramicnetwork/stream-tile-handler@2.2.5-rc.0...@ceramicnetwork/stream-tile-handler@2.2.5-rc.1) (2022-08-11)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
+## [2.2.5-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@2.2.4...@ceramicnetwork/stream-tile-handler@2.2.5-rc.0) (2022-08-10)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
+## [2.2.4](/compare/@ceramicnetwork/stream-tile-handler@2.2.4-rc.2...@ceramicnetwork/stream-tile-handler@2.2.4) (2022-08-08)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
 ## [2.2.4-rc.2](/compare/@ceramicnetwork/stream-tile-handler@2.2.4-rc.1...@ceramicnetwork/stream-tile-handler@2.2.4-rc.2) (2022-07-26)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile-handler